requirements.txt
Pillow==9.0.1
black==21.5b1
comet-ml==3.0.0
cython==0.29.*
dipy==1.5.*
flake8-docstrings==1.6.0
flake8==3.9.2
fury
h5py==2.10.*
holoviews
isort==5.8.0
matplotlib>=2.2.*
nibabel==3.0.*
nilearn
noise
numpy==1.21.*
pandas
pathos
plotly
pre-commit>=2.9.0
ptitprince
pybids
pydot
pygame
pytest-cov
pytest-pep8
pytest-xdist
pytest==5.3.*
pyyaml
scikit-image
scikit-learn==0.22.*
scipy==1.4.*
seaborn==0.11.*
torchsummary==1.5.1
tqdm
trimeshpy
umap-learn
vtk>=9.1.*
xlrd==1.2.0
xlsxwriter==1.3.8
